pyproject.toml
[project]
name = "admin"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "firebase-admin>=7.3.0",
    "flask>=3.1.3",
    "gunicorn>=25.3.0",
    "python-dotenv>=1.2.2",
]
